react query and front bugs fixed for now
